requirements/base.txt
Django==1.8.4
django-bootstrap3==6.2.2
django-cookie-law==1.0.6
logilab-common==1.0.2
Pillow==3.0.0
pyyaml==3.11
six==1.9.0
unipath==1.1
wheel==0.24.0
djangorestframework==3.3.3
